+++ src/tree_grid/model/tests.rs
@@ -4,7 +4,7 @@ use ndarray::Array2;
 
 use super::*;
 
-fn setup_data() -> (Array2<f32>, Array1<f32>) {
+fn setup_data() -> (Array2<f64>, Array1<f64>) {
     let mut rdr = ReaderBuilder::new()
         .has_headers(true)
         .from_path("src/tree_grid/model/dat.csv")
@@ -15,9 +15,9 @@ fn setup_data() -> (Array2<f32>, Array1<f32>) {
 
     for result in rdr.records() {
         let record = result.expect("Failed to read record");
-        let y: f32 = record[0].parse().expect("Failed to parse y");
-        let x1: f32 = record[1].parse().expect("Failed to parse x1");
-        let x2: f32 = record[2].parse().expect("Failed to parse x2");
+        let y: f64 = record[0].parse().expect("Failed to parse y");
+        let x1: f64 = record[1].parse().expect("Failed to parse x1");
+        let x2: f64 = record[2].parse().expect("Failed to parse x2");
 
         y_data.push(y);
         x_data.push(vec![x1, x2]);
